+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ---
 
 
+1.1.0
+
+- changed: button range to [1-9] instead of [1-4] (some user want to use 1-x for something else and use the rest)
+- added: user preferences for the adodn with operator to disable/enable button sets.
+
 1.0.0
 
 - added: Quick selection set in pose bone